User simplifications.
