+++ kansha/models.py
@@ -23,18 +23,3 @@ from kansha import pickle
 
 class Entity(EntityBase, pickle.UnpicklableMixin):
     __metaclass__ = EntityMeta
-
-
-class DataToken(Entity):
-    using_options(tablename='token')
-    token = Field(Unicode(255), unique=True, nullable=False, primary_key=True)
-    username = Field(Unicode(255), nullable=True)
-    date = Field(DateTime, nullable=False)
-    action = Field(Unicode(255), nullable=True)
-    board = ManyToOne('DataBoard')
-
-    @classmethod
-    def delete_token_by_username(cls, username, action):
-        token = cls.get_by(username=username, action=action)
-        if token:
-            token.delete()
