+++ kansha/user/usermanager.py
@@ -7,35 +7,25 @@
 # the file LICENSE.txt, which you should have received as part of
 # this distribution.
 #--
-
-from .models import DataUser
-from nagare.namespaces import xhtml
-from nagare import component, i18n
-from ..toolbox import autocomplete
 import random
 from datetime import datetime, timedelta
 
+from nagare.namespaces import xhtml
+from nagare import component, i18n
 
-def get_user_class(source):
-    """ Return User Class for a given source
-
-    Generic method (from peak). Authentication system implements this
-    method.
-
-    In:
-     - ``source`` -- login source (i.e application, google...)
-    Return:
-     - the user class
-    """
-    raise Exception("User class for source %s not found" % source)
+from ..toolbox import autocomplete
+from .models import DataUser
+from .comp import User
 
 
 def get_app_user(username, data=None):
     """Return User instance"""
     if not data:
         data = UserManager.get_by_username(username)
-    klass = get_user_class(data.source)
-    return klass(username, data=data)
+    if data.source != 'application':
+        # we need to set a passwd for nagare auth
+        return User(username, 'passwd', data=data)
+    return User(username, data=data)
 
 
 class UserManager(object):
@@ -47,7 +37,7 @@ class UserManager(object):
 
     @staticmethod
     def get_all_users(hours=0):
-        """Return all users if `hours` is 0 or just those who have registrated
+        """Return all data users if `hours` is 0 or just those who have registrated
         for the last hours."""
         q = DataUser.query
         if hours:
@@ -135,6 +125,8 @@ class UserManager(object):
                                                                           creation_date=datetime.utcnow(),
                                                                           author=u2))
 
+
+###### TODO: Move the defintions below somewhere else ##########
 
 class NewMember(object):
